Implement HTTPPUT /consume DecreaseQuantity()
+ Update Ingredient model (quantity double => int)

+++ InventoryAPI/Migrations/20210124070843_Init.cs
@@ -14,7 +14,7 @@ namespace InventoryAPI.Migrations
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: true),
                     UnitPrice = table.Column<decimal>(type: "decimal(18,2)", nullable: false),
-                    QuantityOnStock = table.Column<double>(type: "float", nullable: false)
+                    QuantityOnStock = table.Column<int>(type: "int", nullable: false)
                 },
                 constraints: table =>
                 {
